% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csdynpred_sens.R
\name{csdypred_sens}
\alias{csdypred_sens}
\title{A dynamic prediction for cancer progression based on MCICJM also incorporating sensitivity}
\usage{
csdypred_sens(
  sens_fit = 0.6,
  sens_pred = NULL,
  model = NULL,
  t_biopsies = NULL,
  t_start = 0,
  t_visits = NULL,
  n.step = 50L,
  iter = 250L,
  n.adapt = 150L,
  seed = 100L,
  idVar = "CISNET_ID",
  varname = list(longitimeVar = NULL, longi.bsVar = NULL, surv.bsVar = NULL),
  newdata,
  sigma.inits = 0.27,
  sigMat.inits = NULL,
  qp = 15
)
}
\arguments{
\item{sens_fit}{the sensitivity parameter assumed in the MCICJM fitted on the training (PASS) data. The default value is 0.6.}

\item{sens_pred}{the fixed sensitivity value used in the prediction. If not specified, will automatically extract the value from posterior distribution of the sens parameter from the fitted model.}

\item{model}{the self-fitted MCICJM model. If not specified, the default fitted MCICJMs (according to sensitivity) is used.}

\item{t_biopsies}{the times of prior biopsies (excluding the one at year 0, the start of active surveillance).}

\item{t_start}{the start time point (in years) to predict, e.g., usually the last biopsy. The default value is 0.}

\item{n.step}{the interval of each point estimates. The default value is 50.}

\item{iter}{iteration number of the MCMC sampling after adaption. The default value is 250.}

\item{n.adapt}{iteration number of MCMC sampling in the adaption phase. The default value is 150.}

\item{seed}{seed value for reproducibility The default value is 100.}

\item{idVar}{the name of the ID variable in the testset. The default is "CISNET_ID".}

\item{varname}{the names of related variables, default is the same in the PASS data: \cr
\itemize{
  \item \code{longitimeVar} - the variable name of PSA measurement times
  \item \code{longi.bsVar} - the variable name of baseline covariate (i.e., age centered around 62) in the longitudinal submodel
  \item \code{surv.bsVar} - the variable name of baseline covariate (i.e., PSA density in log scale) in the survival submodel
}}

\item{newdata}{the newdata contains the PSA measurements and baseline covariates of only one patient.}

\item{sigma.inits}{the initial value of \mjeqn{\sigma}{ASCII representation} in Robbins-Monro process. The default value is 0.27.}

\item{sigMat.inits}{the initial value of \mjeqn{\Sigma}{ASCII representation} in Robbins-Monro process.}

\item{qp}{the number of Gaussian quadrature point in the numerical approximation for the integral. The default value is 15.}
}
\value{
a list of following components: \cr
\loadmathjax
\itemize{
 \item \code{risk_visits} - the table of predicted risks at the clinical visits of interest, containing four columns: \cr
 \loadmathjax
 \itemize{
   \item \code{t.hor} - time point
   \item \code{mean} - mean estimates
   \item \code{lower} - lower bound of the 95\% credible interval
   \item \code{upper} - lower bound of the 95\% credible interval.
 } \cr
 \item \code{risk} - the risk estimates (with uncertainty) of all time points. This can be used to draw a risk curve.
 \item \code{priorbiopsy_info} - a list includes two tables:
 \loadmathjax
 \itemize{
   \item \code{priorbsp.risk} - risk estimates for each prior biopsy intervals (rows = iterations, cols = biopsy intervals)
   \item \code{baseline.bp.risk} - the offsets of the risks attributed from the biopsy sensitivity in all iterations
 } \cr
 \item \code{longi} - the estimated PSA trajectory (with uncertainty).
 \item \code{acc} - the acceptance rate in the Robbins-Monro process. This is used to check convergence.
 \item \code{surv} - the vector of survival probability until \code{t_start}.
 \item \code{parameters} - the MCMC samples of the parameters from the MCICJM model: \cr
 \loadmathjax
 \itemize{
   \item \code{beta} - fixed effects in the longitudinal submodel
   \item \code{b} - patient-specific random effects in the longitudinal submodel
   \item \code{gam_bh} - the coefficients in the baseline hazard
   \item \code{gamma} - the coefficient of the baseline covariate (PSA density) in the survival submodel
   \item \code{alpha} - the association parameters of PSA in different functional forms
   \item \code{Sigma} - the variance covariance matrix of the random effects in the longitudinal submodel
   \item \code{tau} - the inverse of the residuals variance in the longitudinal submodel
   \item \code{sens} - the sensitivity parameter used in each iteration, either fixed or sampled from posterior
 } \cr
 \item \code{inits} - the initial values for \mjeqn{\sigma}{ASCII representation} and \mjeqn{\Sigma}{ASCII representation} matrix in the Robbins-Monro process.
 \item \code{overall.surv} - a table of the estimates overall survival probabilities at all time points.
 \item \code{full.results} - a list of calculations from MCMC samples at each time points.
}
}
\description{
This function allows you to predict patient-specific risk of cancer progression based on the fitted MCICJM given a certain sensitivity value (or a sensitivity prior). \cr
The risk refers to \loadmathjax \mjdeqn{\Pr(T^*_i = T^{\small{prg},*}_i < t^{(p)} \mid T^{\small{prg}+}_i > t^{(b)})}{ASCII representation}
The prediction can be dynamically updated with new PSA values.
}
\details{
The prediction of cancer progression are based on seven MCICJMs fitted on the Canary PASS data with fixed biopsy sensitivity ranging from 0.6 to 0.9 (with a step of 0.05). According to the user-specified sensitivity in the above-mentioned range, the corresponding MCICJM is chosen as the training model. \cr
This function can only generate predictions based on a dataset containing only ONE patient. This facilitates the users to make prediction in parallel (e.g., using packages \code{doParallel} or \code{future}) if they have larger datasets.
}
\examples{
csdypred_sens(sens_pred = 0.8, model = mcicjm, t_biopsy = 1:2, t_start = 2, t_visits = c(1, 2, 4, 6), newdata = test)
}
\keyword{Dynamic}
\keyword{prediction}
