% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcicjmsim.R
\name{mcicjmsim}
\alias{mcicjmsim}
\title{A simulation function based on MCICJM fitted on PASS data}
\usage{
mcicjmsim(
  n = 1000,
  seed = 100,
  param_list = list(t.max = 12.48, mean.Cens = 5.114849, knot.longi = c(0, 1.36, 2.96,
    12.48), knot.surv = c(0, 0, 0, 0, 1.386667, 2.773333, 4.16, 5.546667, 6.933333, 8.32,
    9.706667, 11.093333, 12.48, 12.48, 12.48, 12.48), age.mean = -0.1248499, age.sd =
    6.86713, density.mean = -2.272663, density.sd = 0.6042731, cvisit.sep = 0.25,
    cvisit.sd = 0.036),
  sens_fit = 0.6,
  sens_sim = NULL,
  sched_sim = "PASS",
  selfdefsched_code = NULL,
  Pcomp = 1,
  Bcomp = 1,
  keep_complete = F
)
}
\arguments{
\item{n}{number of patients on active surveillance.}

\item{seed}{seed value for reproducibility The default value is 100.}

\item{param_list}{the estimated parameters from the PASS data and the fitted MCICJM: \cr
\itemize{
  \item \code{t.max} - maximum follow-up time in the PASS data
  \item \code{mean.Cens} - mean censoring time in the PASS data
  \item \code{knot.longi} - knots used in the natural cubic spline specification of the longitudinal model
  \item \code{knot.surv} - knots used in the P-splines of the baseline hazard in the survival model
  \item \code{age.mean} - mean age observed in the PASS data (centered around 62)
  \item \code{age.sd} - standard deviation of age observed in the PASS data
  \item \code{density.mean} - mean of the PSA density (in log) observed in the PASS data
  \item \code{density.sd} - standard deviation of the PSA density (in log) observed in the PASS data
  \item \code{cvisit.sep} - regular clinical visit interval for PSA measurement. The default is 3 months (0.25 years)
  \item \code{cvisit.sd} - standard deviation of variation in the clinical visit time
}}

\item{sens_fit}{biopsy sensitivity used to generate the underlying event times. This value can be chose from (0.6, 0.65, 0.7, 0.75, 0.8, 0.85, 0.9).}

\item{sens_sim}{biopsy sensitivity used to generate the biopsy time in the simulated dataset. The default value is the same as `sens_fit`.}

\item{sched_sim}{the biopsy schedule implemented in the simulated dataset, can be chosen from \{"PASS" (PASS protocol), "yearly" (yearly biopsies), "biyearly" (biennial biopsies), "triyearly" (triennial biopsies)\}.}

\item{Pcomp}{the compliance rate of PSA measurements. The default value is 1.}

\item{Bcomp}{the compliance rate of biopsies. The default value is 1.}

\item{keep_complete}{also keep the complete datasets (PSA after event occurrence).}
}
\value{
three datasets. First dataset records all the longitudinal measurements (used in the longitudinal submodel), second records each subject per row (used in the survival submodel). Each datasets contains the following columns: \cr
\loadmathjax
\itemize{
 \item \code{CISNET_ID} - patient ID
 \item \code{TimeSince_Dx} - PSA time since start of AS
 \item \code{PSAValue} - transformation of PSA value (ng/ml), \mjeqn{\log_2(\text{PSA} + 1)}{ASCII representation}
 \item \code{time} - true event time in practice
 \item \code{time.cmp1} - progression-free time
 \item \code{time.cmp2} - treatment-free time
 \item \code{status.cmp} - observed event indicator (1 = cancer progression; 2 = early treatment; 0 = censoring)
 \item \code{density} - baseline PSA density (\mjeqn{\text{ng}/\text{ml}^2}{ASCII representation}) in log transformation, \mjeqn{\log(\text{PSA density})}{ASCII representation}
 \item \code{DxAge} - baseline age, centered around 62 years old
 \item \code{b1 - b4} - true random effects for PSA
 \item \code{time.prg} - true cancer progression time.
 \item \code{time.trt} - true early treatment time.
 \item \code{time.cen} - true censoring time.
} \cr
The third dataset contains all the biopsies each patient participated in and has the following columns: \cr
\loadmathjax
\itemize{
 \item \code{CISNET_ID} - patient ID
 \item \code{TimeSince_Dx} - biopsy time since start of AS
 \item \code{compliance} - whether the patient participates in this biopsy
 \item \code{bioresult} - whether the patient is detected with progression
 \item \code{status.cmp} - observed event indicator (1 = cancer progression; 2 = early treatment; 0 = censoring)
 \item \code{time} - true event time in practice
 \item \code{time.cmp2} - treatment-free time, which is also the time until which the patient is followed
}
}
\description{
This function allows you to simulate the outcomes (censoring, progression, early treatment) of patients in AS based on PSA value and baseline age and PSA density. \cr
In addition to \code{icjmsim()}, this function considers the biopsy sensitivity (0.6 to 0.9 with a step pf 0.05).
}
\details{
The simulated data are based on seven MCICJMs fitted on the Canary PASS data with fixed biopsy sensitivity ranging from 0.6 to 0.9 (with a step of 0.05). According to the user-specified sensitivity in the above-mentioned range, the following parameters are corresponding chosen \cr
\loadmathjax
\itemize{
  \item \code{betas} - coefficients for the fixed effects in the longitudinal submodel for PSA
  \item \code{sigma.y} - the standard deviation for the residuals in the longitudinal submodel for PSA
  \item \code{D_c3} - the covariance matrix for the random effects in the longitudinal submodel for PSA
  \item \code{gambh} - coefficients for the P-spline design matrix of the baseline hazard in the survival submodel
  \item \code{gammas} - coefficients for the exogenous covariates in the survival submodel;
  \item \code{alpha} - coefficients for the impact from PSA on the time-to-event outcomes in the survival submodel
} \cr
The `true event time` is simulated based on the event-specific hazard and bounded with the observed maximum follow-up time (12.48 yrs) in the PASS data. The `censoring time` follows a uniform distribution with the upper bound as the doubled mean of the observed censoring time in the PASS data. \cr
This function also allows to simulate datasets with newly-specified biopsy sensitivity and different biopsy schedules from the PASS data.
}
\examples{
mcicjmsim(n = 1e5, sens_fit = 0.7)
}
\keyword{MCICJM}
\keyword{Simulation}
